Remove already reported crash reports on android (#29)
On Android, we use
`activityManager.getHistoricalProcessExitReasons(...)` to get the
latests crash reports.
These are stored on device and are persisted. I think it keeps the last
100 crash reports.

We were asking android for these reports on each start of our package.
So we kept getting the same reports.

To fix this. I now store the handled reports in sharedPrefs. And then
ignore these reports when I grab the next report.