feat: add header to config to send with each request (#59)
## Summary

This pull request adds support for custom HTTP headers in the
`FaroConfig`, allowing users to specify headers that will be included in
all outgoing requests made by the SDK.

## Context

In our internal usage of the faro-flutter-sdk, requests to our Alloy
instance were returning 404 Not Found responses. After debugging, we
identified that our deployment requires a host header to correctly route
and process tracking events.

To support this, we introduced a configurable `headers` field in
`FaroConfig`, which enables us (and others with similar requirements) to
inject any necessary headers globally.

## Changes

- Added a new `Map<String, String>? headers` field to FaroConfig.
- These headers are now included in all HTTP requests made by the SDK.