feat: implement persistent data collection setting across app restarts (#64)
## Description

Before the `enableDataCollection` settings was only kept in memory. So
if you toggled it off, then on next app start, it would be on again.
This fix changes this behavior so that the setting is persisted between
app starts.
The setting is stored in UserDefaults on iOS and SharedPreferences on
Android.

Also updates the example app with a new button to toggle the
enableDataCollection and show that current state.

## Related Issue(s)

Fixes #62 

## Type of Change

- [x] 🛠️ Bug fix (non-breaking change which fixes an issue)
- [ ] 🚀 New feature (non-breaking change which adds functionality)
- [ ] 💥 Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] 📝 Documentation
- [ ] 📈 Performance improvement
- [ ] 🏗️ Code refactoring
- [ ] 🧹 Chore / Housekeeping

## Checklist

- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the CHANGELOG.md under the "Unreleased" section