feat(tracing): enhance span status management (#87)
## Description

Add a new boolean property `statusHasBeenSet` to the `Span` interface to
track if the span status has already been set. Update the
`FaroZoneSpanManager` to conditionally set the span status to OK or
ERROR based on this property.
Introduce comprehensive tests to verify the new behavior, ensuring that
the status is only set when it hasn't been previously defined.

## Related Issue(s)

Fixes #86 

## Type of Change

- [x] 🛠️ Bug fix (non-breaking change which fixes an issue)
- [ ] 🚀 New feature (non-breaking change which adds functionality)
- [ ] 💥 Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 📝 Documentation
- [ ] 📈 Performance improvement
- [ ] 🏗️ Code refactoring
- [ ] 🧹 Chore / Housekeeping

## Checklist

- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the CHANGELOG.md under the "Unreleased" section